修改源码类型定义错误问题

+++ msg.h
@@ -9,7 +9,7 @@ uint16_t get_num_msg(struct msg_data *msg);
 void get_msg_boundaries(struct msg_data *msg, uint16_t msg_num, uint16_t *start_page, uint16_t *end_page);
 uint16_t find_last_msg(struct msg_data *msg);
 uint16_t find_last_page(struct msg_data *msg);
-uint16_t find_last_page_of_msg(struct msg_data *msg, uint16_t msg_number);
+int find_last_page_of_msg(struct msg_data *msg, uint16_t msg_number);
 int check_wrapped(struct msg_data *msg);
 
 #endif
